requirements.txt
requests
mysql-connector-python
pandas
matplotlib
scipy
mysqlclient